+++ seleniumide/src/main/java/org/musetest/seleniumide/values/JavascriptValueConverter.java
@@ -1,25 +1,23 @@
 package org.musetest.seleniumide.values;
 
-import org.musetest.builtins.value.*;
-import org.musetest.core.values.*;
-import org.musetest.seleniumide.*;
+import org.musetest.javascript.*;
 
 /**
  * @author Christopher L Merrill (see LICENSE.txt for license details)
  */
 @SuppressWarnings("unused") // invoked via reflection
-public class VariableValueConverter extends PrefixedBracketedValueConverter
+public class JavascriptValueConverter extends PrefixedBracketedValueConverter
     {
     @Override
     protected String getValueSourceTypeId()
         {
-        return VariableValueSource.TYPE_ID;
+        return EvaluateJavascriptValueSource.TYPE_ID;
         }
 
     @Override
     protected String getPrefix()
         {
-        return "$";
+        return "javascript";
         }
     }